Fixes #16 Cross Channel Chatter
If a user would abandon a request due to the
mis-behavior of the selected gateway node,
communication errors from the previous request
would bleed into the subsequent request results.

This was due to re-use of components by the
blazor framework, added a reference count for
requests to ignore results from old requests.

In the future, one may consider adding a
cancellation token to the .net sdk to
explicitly terminate such requests.